Reorganize code, and simplify some duplication
